♻️ 불필요한 텔레포트 파일 삭제
사용하지 않는 모든 텔레포트 코드 삭제
