+++ Visiom/FullImmersive/FullImmersiveView+handlers.swift
@@ -11,34 +11,6 @@ import SwiftUI
 
 // MARK: - Handlers Extension
 extension FullImmersiveView {
-//
-//    var teleportDragWaypoint: some Gesture {
-//        DragGesture()
-//            .targetedToAnyEntity()
-//            .onChanged { value in
-//                value.entity.position = value.convert(value.location3D, from: .local, to: value.entity.parent!)
-//            }
-//    }
-    
-    func tapToTeleport(value: EntityTargetValue<TapGesture.Value>) {
-        guard let sceneContent = self.root else { return }
-        
-        // Calculate the vector from the origin to the tapped position
-        let vectorToTap = value.entity.position
-        
-        // Normalize the vector to get a direction from the origin to the tapped position
-        let direction = normalize(vectorToTap)
-        
-        // Calculate the distance (or magnitude) between the origin and the tapped position
-        let distance = length(vectorToTap)
-        
-        // Calculate the new position by inverting the direction multiplied by the distance
-        let newPosition = -direction * distance
-        
-        // Update sceneOffset's X and Z components, leave Y as it is
-        sceneContent.position.x = newPosition.x
-        sceneContent.position.z = newPosition.z
-    }
     
     /// 엔티티 계층 구조 업데이트
     func updateEntityHierarchy() {
